requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dateutil==2.8.2
pytest==7.4.3
pytest-flask==1.3.0
Flask-Migrate==4.0.5
gunicorn==21.2.0